+++ webapp/mydata.json
@@ -1,7 +1,12 @@
 { 
-    "oModel": 
-    {   
-        "1": "Case 1: The Adventure of The Unholy Man",
-        "2": "Case 2: "  
-    }
+    "oModel": [
+        {   
+            "id": "1",
+            "desc": "Case 1: Adventures of the Unholy Man"  
+        },
+        {
+            "id": "2",
+            "desc": "Case 2: ..."
+        }
+    ]
 }
