+++ webapp/view/MainPanel.view.xml
@@ -9,7 +9,7 @@
         <content>
             <Button
                 text="{i18n>showHelloButtonText}"
-                press=".onShowHello"
+                press=".onShowData"
                 class="sapUiSmallMarginEnd"/>
             <!-- provided the below an ID so I can retrieve its value -->
             <InputBase
@@ -25,7 +25,7 @@
         <content>
             <Button
                 text="{i18n>showHelloButtonText}"
-                press=".onShowHello"
+                press=".onShowData"
                 class="sapUiSmallMarginEnd"/>
             <InputBase
                 id="inputSolution"
@@ -40,7 +40,7 @@
         <content>
             <Button
                 text="{i18n>showHelloButtonText}"
-                press=".onShowHello"
+                press=".onShowData"
                 class="sapUiSmallMarginEnd"/>
             <InputBase
                 id="inputHint"
